Move devices under com.oktawski.iotserver.devices
