+++ src/main/java/com/oktawski/iotserver/devices/light/LightController.java
@@ -1,4 +1,4 @@
-package com.oktawski.iotserver.light;
+package com.oktawski.iotserver.devices.light;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
